Move diagnostics management to its own submodule
