+++ esmvalcore/_task.py
@@ -19,7 +19,7 @@ import psutil
 import yaml
 
 from ._citation import _write_citation_files
-from ._config import DIAGNOSTICS_PATH, TAGS, replace_tags
+from ._diagnostics import DIAGNOSTICS_PATH, TAGS, replace_tags
 from ._provenance import TrackedFile, get_task_provenance
 
 logger = logging.getLogger(__name__)
